requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
torch
pypdf
numpy
